FoodRecognize_FastAPI/app/requirement.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
Pillow
json
os
quaterion
numpy